Build cgal comm and start moving visibility graph.
