+++ polygon_coverage_geometry/CMakeLists.txt
@@ -21,6 +21,7 @@ cs_add_library(${PROJECT_NAME}
   #src/geometry/polygon.cc
   #src/geometry/BCD.cpp
   src/bcd.cc
+  src/cgal_comm.cc
   src/weakly_monotone.cc
   #src/geometry/sweep.cc
 )
