+++ polygon_coverage_geometry/include/polygon_coverage_geometry/cgal_definitions.h
@@ -14,11 +14,14 @@ typedef K::Line_2 Line_2;
 typedef K::Intersect_2 Intersect_2;
 typedef K::Plane_3 Plane_3;
 typedef K::Segment_2 Segment_2;
+typedef K::Triangle_2 Triangle_2;
 typedef CGAL::Polygon_2<K> Polygon_2;
 typedef Polygon_2::Vertex_const_iterator VertexConstIterator;
 typedef Polygon_2::Vertex_const_circulator VertexConstCirculator;
 typedef Polygon_2::Edge_const_iterator EdgeConstIterator;
 typedef Polygon_2::Edge_const_circulator EdgeConstCirculator;
 typedef CGAL::Polygon_with_holes_2<K> PolygonWithHoles;
+
+enum DecompositionType { kBoustrophedeon, kTrapezoidal };
 
 #endif  // POLYGON_COVERAGE_GEOMETRY_CGAL_DEFINITIONS_H_
